Create compas-prediction-age.ipynb
Correlation between prediction and the true labels for the COMPAS study. The dataset is grouped by age groups and protected classes. Gender is marked with solid lines; Race/ ethnicity with dashed lines. With the exception of females which exhibit a slight decline between the last two age groups all other trends are ascending.